pyproject.toml
[project]
name = "pleiades-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for academic paper search, citation formatting, and impact metrics using Google Scholar and LLM elicitation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
 "mcp>=1.1.2",
 "google-search-results>=2.4.0",
 "python-dotenv>=1.0.0",
 "requests>=2.25.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[[project.authors]]
name = "akougkas.io (Anthony K)"
email = "a.kougkas@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pleiades-mcp = "pleiades_mcp:main"
